+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/ide/inspections/compilation/RobotKeywordAnnotator.java
@@ -7,16 +7,13 @@ import com.intellij.lang.annotation.HighlightSeverity;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiReference;
 import com.intellij.psi.util.PsiTreeUtil;
-import com.jetbrains.python.psi.PyElement;
-import com.jetbrains.python.psi.PyElementVisitor;
 import dev.xeonkryptos.xeonrobotframeworkplugin.RobotBundle;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotKeywordCall;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotKeywordCallName;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotLocalSetting;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotStatement;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotTemplateStatementsGlobalSetting;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.ref.PyElementDeprecatedVisitor;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.ref.PyElementParentTraversalVisitor;
+import dev.xeonkryptos.xeonrobotframeworkplugin.util.DeprecationInspector;
 import org.jetbrains.annotations.NotNull;
 
 public class RobotKeywordAnnotator implements Annotator {
@@ -35,19 +32,18 @@ public class RobotKeywordAnnotator implements Annotator {
                   .highlightType(ProblemHighlightType.ERROR)
                   .range(element)
                   .create();
-        } else if (resolvedElement instanceof PyElement pyElement) {
-            PyElementDeprecatedVisitor pyElementDeprecatedVisitor = new PyElementDeprecatedVisitor();
-            PyElementVisitor pyElementParentTraversalVisitor = new PyElementParentTraversalVisitor(pyElementDeprecatedVisitor);
-            pyElement.accept(pyElementParentTraversalVisitor);
-
-            if (pyElementDeprecatedVisitor.isDeprecated()) {
+        } else {
+            if (DeprecationInspector.isDeprecated(resolvedElement)) {
                 holder.newSilentAnnotation(HighlightSeverity.WARNING).range(element).highlightType(ProblemHighlightType.LIKE_DEPRECATED).create();
             }
         }
 
         RobotKeywordCall keywordCall = PsiTreeUtil.getParentOfType(robotKeywordCallName, RobotKeywordCall.class);
         if (keywordCall != null && !keywordCall.allRequiredParametersArePresent()) {
-            RobotStatement ignoringParameterCheckParent = PsiTreeUtil.getParentOfType(keywordCall, true, RobotLocalSetting.class, RobotTemplateStatementsGlobalSetting.class);
+            RobotStatement ignoringParameterCheckParent = PsiTreeUtil.getParentOfType(keywordCall,
+                                                                                      true,
+                                                                                      RobotLocalSetting.class,
+                                                                                      RobotTemplateStatementsGlobalSetting.class);
             if (ignoringParameterCheckParent == null) {
                 String missingRequiredParameters = String.join(", ", keywordCall.computeMissingRequiredParameters());
                 holder.newAnnotation(HighlightSeverity.ERROR, RobotBundle.getMessage("annotation.keyword.parameters.missing", missingRequiredParameters))