+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/psi/element/DefinedKeyword.java
@@ -25,4 +25,6 @@ public interface DefinedKeyword {
    String getArgumentsDisplayable();
 
    boolean isDeprecated();
+
+    boolean isPrivate();
 }