+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/psi/ref/PyElementDeprecatedVisitor.java
@@ -4,7 +4,11 @@ import com.jetbrains.python.psi.PyClass;
 import com.jetbrains.python.psi.PyDecorator;
 import com.jetbrains.python.psi.PyDecoratorList;
 import com.jetbrains.python.psi.PyElementVisitor;
+import com.jetbrains.python.psi.PyExpressionStatement;
 import com.jetbrains.python.psi.PyFunction;
+import com.jetbrains.python.psi.PyStatementList;
+import com.jetbrains.python.psi.PyStringLiteralExpression;
+import dev.xeonkryptos.xeonrobotframeworkplugin.util.GlobalConstants;
 import org.jetbrains.annotations.NotNull;
 
 public class PyElementDeprecatedVisitor extends PyElementVisitor {
@@ -13,26 +17,48 @@ public class PyElementDeprecatedVisitor extends PyElementVisitor {
 
     @Override
     public void visitPyClass(@NotNull PyClass node) {
+        super.visitPyClass(node);
         node.acceptChildren(this);
     }
 
     @Override
     public void visitPyFunction(@NotNull PyFunction node) {
+        super.visitPyFunction(node);
         node.acceptChildren(this);
     }
 
     @Override
     public void visitPyDecoratorList(@NotNull PyDecoratorList node) {
+        super.visitPyDecoratorList(node);
         node.acceptChildren(this);
     }
 
     @Override
     @SuppressWarnings("UnstableApiUsage")
     public void visitPyDecorator(@NotNull PyDecorator node) {
+        super.visitPyDecorator(node);
         String decoratorName = node.getName();
-        if ("deprecated".equals(decoratorName)) {
+        if ("deprecated".equalsIgnoreCase(decoratorName)) {
             deprecated = true;
         }
+    }
+
+    @Override
+    public void visitPyStatementList(@NotNull PyStatementList node) {
+        super.visitPyStatementList(node);
+        node.acceptChildren(this);
+    }
+
+    @Override
+    public void visitPyExpressionStatement(@NotNull PyExpressionStatement node) {
+        super.visitPyExpressionStatement(node);
+        node.acceptChildren(this);
+    }
+
+    @Override
+    public void visitPyStringLiteralExpression(@NotNull PyStringLiteralExpression node) {
+        super.visitPyStringLiteralExpression(node);
+        deprecated |= node.getText().startsWith(GlobalConstants.DEPRECATED_PREFIX);
     }
 
     public boolean isDeprecated() {