+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/util/GlobalConstants.java
@@ -9,6 +9,8 @@ public final class GlobalConstants {
 
     public static final String ROBOT_BUILT_IN = "robot.libraries.BuiltIn";
 
+    public static final String DEPRECATED_PREFIX = "*DEPRECATED";
+
     private GlobalConstants() {
         throw new UnsupportedOperationException("This is a utility class and cannot be instantiated");
     }